+++ pipeline/tick/alert_test.go
@@ -314,6 +314,7 @@ func TestAlertPagerDuty2(t *testing.T) {
 	pipe, _, from := StreamFrom()
 	handler := from.Alert().PagerDuty2()
 	handler.RoutingKey = "LeafsNation"
+	handler.Link("https://example.com/chart", "some chart")
 
 	want := `stream
     |from()
@@ -324,6 +325,49 @@ func TestAlertPagerDuty2(t *testing.T) {
         .history(21)
         .pagerDuty2()
         .routingKey('LeafsNation')
+        .link('https://example.com/chart', 'some chart')
+`
+	PipelineTickTestHelper(t, pipe, want)
+}
+
+func TestAlertPagerDuty2MissingLinkText(t *testing.T) {
+	pipe, _, from := StreamFrom()
+	handler := from.Alert().PagerDuty2()
+	handler.RoutingKey = "LeafsNation"
+	handler.Link("https://example.com/chart")
+
+	want := `stream
+    |from()
+    |alert()
+        .id('{{ .Name }}:{{ .Group }}')
+        .message('{{ .ID }} is {{ .Level }}')
+        .details('{{ json . }}')
+        .history(21)
+        .pagerDuty2()
+        .routingKey('LeafsNation')
+        .link('https://example.com/chart')
+`
+	PipelineTickTestHelper(t, pipe, want)
+}
+
+func TestAlertPagerDuty2MultipleLinks(t *testing.T) {
+	pipe, _, from := StreamFrom()
+	handler := from.Alert().PagerDuty2()
+	handler.RoutingKey = "LeafsNation"
+	handler.Link("https://example.com/chart", "some chart")
+	handler.Link("https://example.com/details", "details")
+
+	want := `stream
+    |from()
+    |alert()
+        .id('{{ .Name }}:{{ .Group }}')
+        .message('{{ .ID }} is {{ .Level }}')
+        .details('{{ json . }}')
+        .history(21)
+        .pagerDuty2()
+        .routingKey('LeafsNation')
+        .link('https://example.com/chart', 'some chart')
+        .link('https://example.com/details', 'details')
 `
 	PipelineTickTestHelper(t, pipe, want)
 }